+++ src/main/resources/db/migration/V5__Create_comment_table.sql
@@ -3,7 +3,7 @@ create table COMMENT
     ID           BIGINT auto_increment,
     PARENT_ID    BIGINT not null,
     TYPE         INT    not null,
-    COMMENTATOR  INT    not null,
+    COMMENTATOR  BIGINT    not null,
     GMT_CREATE   BIGINT not null,
     GMT_MODIFIED BIGINT not null,
     LIKE_COUNT   BIGINT default 0,
